Updates Heapsort - tests - main
Heapsort updated so the points are sorted based on the angle of the line
defined by them and the vertex with min-y coordinate.
Tests updated accordingly + written a test for Utils::isCounterClockwise function
Refactored main so the convexHull algorithm is on a separate function
and the points are standard and trivial, so debugging gets easier. (Will need to change on future commit)