Adds OpenGL namespace
Implements a function that will convert vectors of pointers to Vertex instances
to an array.
A function that draws the final graph, using the Vertices convexHull returns,
the initial points (on which Convex Hull run), using the mapper and OpenGL
Functions that will initialize and terminate a window, using GLFW3.