Fixes a bug on AVLTree, Implements helping methods
BUG: Deleting root with two children -> deleting successor ->
tree rebalances changing the root of the tree -> I need to replace the
child of the new parent (of the previous root) with the successor,
but parent is not available.
CHOICES:
    1. Search for the parent of the root, if one (should handle the case
    root does not rotate) - messy
    2. Replace the left and right child of the successor node with the
    children of the deleted node (prev known as root) as well as the
    data.
FIX: The second choice
TEST: A test case was introduced with some comments on how to improve
them
NEW: Implements:
    - insertArray and deleteArray to make some batch inserts/
deletions with less code.
    - getPredecessorData and getSuccessorData so I do not expose the
    wrapper class node
    - getNextSmallestData(data) and getNextLargestData(data),
    so I may find the predecessor and successor data of a value that is
    not inside the tree

TODO: Choose to expose certain methods from the AVLTree, as well as
merging some of the existing ones using wrapper methods that will make
them safe (ex. getSuccessorData(data) should return the correct value
event if data is not in the tree - using getNextLargestData)