Introduces env - Updates Segment comparisson data
- Why introduce env, that contain constants about precision?
Those constants will define how precise we require our Points to be and
thus minimizing the errors that occure when performing operations on segments
(continuous) to find intersection points that we may only represent
with limited percision.

- Segment comparisson data
A Segment should be equal to another segment if both Points are equal,
but since we also need to store them in an AVLTree that requires them to
have an order, we need to specify when a Segment is considered to be
greater or smaller that another Segment. In our algorithm a segment
is considered to be smallest in the status Tree, if it is the first
segment to be intersected by the sweep line. Thus we require a way to
calculate the point where the sweepline intersects the Segment -
updateComparissonData() method.