OpenGL, pickle, fine tuning on the algorithm
The visual representation of the segments and the intersection points
found by the sweep line algorithm, using OpenGL
It exposed some more bugs on the code that I tried to correct, but there
are still intersection points that are not being found
Implements brute force, so I may create scenarios which I know my
implementation does not work, so I may use them for debugging.
In order to save those segments I used the pickle library to dump them
in a .txt file.